refactor: ТЗ-12 8nd save + big changes

+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/FilmDbStorage.java
@@ -2,14 +2,18 @@ package ru.yandex.practicum.filmorate.dao.impl;
 
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.dao.FilmStorage;
+import ru.yandex.practicum.filmorate.exeption.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.MpaRating;
+import ru.yandex.practicum.filmorate.service.UserService;
 
 import java.sql.Date;
 import java.sql.PreparedStatement;
@@ -23,6 +27,7 @@ import java.util.List;
 public class FilmDbStorage implements FilmStorage {
 
     private final JdbcTemplate jdbcTemplate;
+    private UserService userService;
 
     @Override
     public List<Film> listFilms() {
@@ -32,8 +37,12 @@ public class FilmDbStorage implements FilmStorage {
 
     @Override
     public Film getFilmById(int id) {
-        String sqlQuery = "SELECT * FROM films WHERE film_id = ?";
-        return jdbcTemplate.queryForObject(sqlQuery, this::mapRowToFilm, id);
+        try {
+            String sqlQuery = "SELECT * FROM films WHERE film_id = ?";
+            return jdbcTemplate.queryForObject(sqlQuery, this::mapRowToFilm, id);
+        } catch (EmptyResultDataAccessException e) {
+            throw new NotFoundException("Ошибка запроса фильма, проверьте корректность данных.");
+        }
     }
 
     @Override
@@ -58,43 +67,67 @@ public class FilmDbStorage implements FilmStorage {
 
     @Override
     public Film updateFilm(Film film) {
+        String sqlDeleteFilmGenre = "DELETE FROM film_genres WHERE film_id = ?";
         String sqlQuery = "UPDATE films SET name = ?, description = ?, release_date = ?, duration = ?, " +
                 "rating_id = ? WHERE film_id = ?";
+        String genreSqlQuery = "INSERT INTO film_genres (film_id, id) VALUES (?, ?)";
+        jdbcTemplate.update(sqlDeleteFilmGenre, film.getId());
         jdbcTemplate.update(sqlQuery, film.getName(), film.getDescription(),
                 Date.valueOf(film.getReleaseDate()), film.getDuration(), film.getMpa().getId(), film.getId());
+        film.getGenres().forEach(x -> jdbcTemplate.update(genreSqlQuery, film.getId(), x.getId()));
         return film;
     }
 
     @Override
     public void addLike(int id, int userId) {
-        String sqlQuery = "INSERT INTO likes (film_id, user_id) VALUES (?, ?)";
-        jdbcTemplate.update(sqlQuery, id, userId);
+        if (getFilmById(id).getId() == id && userService.getUserById(userId).getId() == userId) {
+            String sqlQuery = "INSERT INTO likes (film_id, user_id) VALUES (?, ?)";
+            jdbcTemplate.update(sqlQuery, id, userId);
+        } else {
+            log.warn("Ошибка при добавлении лайка фильму.");
+            throw new NotFoundException("Ошибка добавления лайка, проверьте корректность данных.");
+        }
     }
 
     @Override
     public void deleteLike(int id, int userId) {
-        String sqlQuery = "DELETE FROM likes WHERE film_id = ? AND user_id = ?";
-        jdbcTemplate.update(sqlQuery, id, userId);
+        if (getFilmById(id).getId() == id && userService.getUserById(userId).getId() == userId) {
+            String sqlQuery = "DELETE FROM likes WHERE film_id = ? AND user_id = ?";
+            jdbcTemplate.update(sqlQuery, id, userId);
+        } else {
+            log.warn("Ошибка при добавлении лайка фильму.");
+            throw new NotFoundException("Ошибка добавления лайка, проверьте корректность данных.");
+        }
     }
 
     @Override
     public List<Film> getMostPopularFilms(int count) {
         String sqlQuery = "SELECT * FROM films f LEFT JOIN (SELECT film_id, COUNT(*) likes_count" +
                 " FROM likes GROUP BY film_id) l ON f.film_id = l.film_id ORDER BY l.likes_count DESC LIMIT ?";
-        return jdbcTemplate.query(sqlQuery, this::mapRowToFilm);
+        return jdbcTemplate.query(sqlQuery, this::mapRowToFilm, count);
     }
 
-    private Film mapRowToFilm(ResultSet resultSet, int rowNum) throws SQLException {
-        Film film = new Film();
-        film.setId(resultSet.getInt("film_id"));
-        film.setName(resultSet.getString("name"));
-        film.setDescription(resultSet.getString("description"));
-        film.setReleaseDate(resultSet.getDate("release_date").toLocalDate());
-        film.setDuration(resultSet.getInt("duration"));
-        int mpaRating = (resultSet.getInt("rating_id"));
-        MpaRating mpa = jdbcTemplate.queryForObject("SELECT * FROM rating WHERE id = ?", new BeanPropertyRowMapper<>(MpaRating.class), mpaRating);
-        film.setMpa(mpa);
-        return film;
+    private Film mapRowToFilm(ResultSet resultSet, int rowNum) {
+        try {
+            Film film = new Film();
+            film.setId(resultSet.getInt("film_id"));
+            film.setName(resultSet.getString("name"));
+            film.setDescription(resultSet.getString("description"));
+            film.setReleaseDate(resultSet.getDate("release_date").toLocalDate());
+            film.setDuration(resultSet.getInt("duration"));
+            int mpaRating = (resultSet.getInt("rating_id"));
+            MpaRating mpa = jdbcTemplate.queryForObject("SELECT * FROM rating WHERE id = ?", new BeanPropertyRowMapper<>(MpaRating.class), mpaRating);
+            film.setMpa(mpa);
+            film.setGenres(getFilmGenres(film.getId()));
+            return film;
+        } catch (EmptyResultDataAccessException | SQLException e) {
+            throw new NotFoundException("Ошибка запроса, проверьте корректность данных.");
+        }
     }
 
+    private List<Genre> getFilmGenres(int id) {
+        String sqlQuery = "SELECT f.id, name FROM film_genres f" +
+                " LEFT JOIN (SELECT * FROM genres) g ON f.id = g.id WHERE film_id = ?";
+        return jdbcTemplate.query(sqlQuery, new BeanPropertyRowMapper<>(Genre.class), id);
+    }
 }
